データ移行用初期化メソッドをDaoのサブクラスに移動
