+++ src/main/java/org/support/project/knowledge/dao/TemplateItemsDao.java
@@ -1,9 +1,11 @@
 package org.support.project.knowledge.dao;
 
+import org.support.project.aop.Aspect;
 import org.support.project.di.Container;
 import org.support.project.di.DI;
 import org.support.project.di.Instance;
 import org.support.project.knowledge.dao.gen.GenTemplateItemsDao;
+import org.support.project.ormapping.common.SQLManager;
 
 /**
  * テンプレートの項目
@@ -13,7 +15,6 @@ public class TemplateItemsDao extends GenTemplateItemsDao {
 	
 	public static final int ITEM_ID_BOOKMARK_URL = 0;
 
-	
 	/** SerialVersion */
 	private static final long serialVersionUID = 1L;
 	/**
@@ -25,6 +26,14 @@ public class TemplateItemsDao extends GenTemplateItemsDao {
 		return Container.getComp(TemplateItemsDao.class);
 	}
 
-
-
+	/**
+	 * データをtruncateする
+	 * 
+	 * @return void
+	 */
+	@Aspect(advice=org.support.project.ormapping.transaction.Transaction.class)
+	public void truncate() {
+		String sql = SQLManager.getInstance().getSql("/org/support/project/knowledge/dao/sql/TemplateItemsDao/TemplateItemsDao_truncate.sql");
+		executeUpdate(sql);
+	}
 }
